Move content of plotting.py
