+++ baybe/recommenders/pure/bayesian/botorch.py
@@ -32,8 +32,8 @@ from baybe.searchspace import (
     SubspaceContinuous,
     SubspaceDiscrete,
 )
+from baybe.utils.conversion import to_string
 from baybe.utils.dataframe import to_tensor
-from baybe.utils.plotting import to_string
 from baybe.utils.sampling_algorithms import (
     DiscreteSamplingMethod,
     sample_numerical_df,
