+++ baybe/recommenders/pure/nonpredictive/clustering.py
@@ -15,7 +15,7 @@ from typing_extensions import override
 
 from baybe.recommenders.pure.nonpredictive.base import NonPredictiveRecommender
 from baybe.searchspace import SearchSpaceType, SubspaceDiscrete
-from baybe.utils.plotting import to_string
+from baybe.utils.conversion import to_string
 
 
 @define
