+++ baybe/searchspace/discrete.py
@@ -26,10 +26,10 @@ from baybe.searchspace.validation import validate_parameter_names, validate_para
 from baybe.serialization import SerialMixin, converter, select_constructor_hook
 from baybe.utils.basic import to_tuple
 from baybe.utils.boolean import eq_dataframe, strtobool
+from baybe.utils.conversion import to_string
 from baybe.utils.dataframe import get_transform_objects, pretty_print_df
 from baybe.utils.memory import bytes_to_human_readable
 from baybe.utils.numerical import DTypeFloatNumpy
-from baybe.utils.plotting import to_string
 
 if TYPE_CHECKING:
     import polars as pl
