+++ baybe/surrogates/bandit.py
@@ -15,7 +15,7 @@ from baybe.priors import BetaPrior
 from baybe.searchspace.core import SearchSpace
 from baybe.surrogates.base import Surrogate
 from baybe.targets.binary import _FAILURE_VALUE_COMP, _SUCCESS_VALUE_COMP
-from baybe.utils.plotting import to_string
+from baybe.utils.conversion import to_string
 from baybe.utils.random import temporary_seed
 
 if TYPE_CHECKING:
