+++ baybe/surrogates/base.py
@@ -32,8 +32,8 @@ from baybe.serialization.core import (
     unstructure_base,
 )
 from baybe.serialization.mixin import SerialMixin
+from baybe.utils.conversion import to_string
 from baybe.utils.dataframe import to_tensor
-from baybe.utils.plotting import to_string
 from baybe.utils.scaling import ColumnTransformer
 
 if TYPE_CHECKING:
