+++ baybe/surrogates/ngboost.py
@@ -11,7 +11,7 @@ from baybe.parameters.base import Parameter
 from baybe.surrogates.base import IndependentGaussianSurrogate
 from baybe.surrogates.utils import batchify_mean_var_prediction, catch_constant_targets
 from baybe.surrogates.validation import get_model_params_validator
-from baybe.utils.plotting import to_string
+from baybe.utils.conversion import to_string
 
 if TYPE_CHECKING:
     from botorch.models.transforms.input import InputTransform
